+++ src/vars.c
@@ -3,6 +3,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <strings.h>
+#include <sys/types.h>
 
 char *get_shell_varp(char *name, char *def_val) {
     symtable_entry_s *entry = get_symtable_entry(name);
@@ -13,11 +14,11 @@ int get_shell_vari(char *name, int def_val) {
     return (int)get_shell_varl(name, def_val);
 }
 
-long get_shell_varl(char *name, int def_val) {
+ssize_t get_shell_varl(char *name, int def_val) {
     symtable_entry_s *entry = get_symtable_entry(name);
     if (entry && entry->val && entry->val[0]) {
         char *strend = NULL;
-        long i = strtol(entry->val, &strend, 10);
+        ssize_t i = strtol(entry->val, &strend, 10);
         if (strend == entry->val || *strend)
             return def_val;
         return i;
